+++ app/actions/index.js
@@ -1,0 +1,13 @@
+import * as Constants from '../constants';
+
+function makeActionCreator(type, ...argNames) {
+  return function act(...args) {
+    const action = { type };
+    argNames.forEach((arg, index) => {
+      action[argNames[index]] = args[index];
+    });
+    return action;
+  };
+}
+
+export const pomoComplete = makeActionCreator(Constants.POMO_COMPLETE);