+++ app/components/Navigation.jsx
@@ -1,0 +1,34 @@
+import React from 'react';
+import Drawer from 'material-ui/Drawer';
+import MenuItem from 'material-ui/MenuItem';
+import RaisedButton from 'material-ui/RaisedButton';
+import AppBar from 'material-ui/AppBar';
+
+export default class Navigation extends React.Component {
+
+  constructor(props) {
+    super(props);
+    this.state = { open: true };
+  }
+
+  handleToggle = () => this.setState({ open: !this.state.open });
+
+  render() {
+    return (
+      <div>
+        <AppBar
+          title="BreakTrack"
+          iconClassNameRight="muidocs-icon-navigation-expand-more"
+        />
+        <RaisedButton
+          label="Toggle Drawer"
+          onTouchTap={this.handleToggle}
+        />
+        <Drawer open={this.state.open}>
+          <MenuItem>Menu Item</MenuItem>
+          <MenuItem>Menu Item 2</MenuItem>
+        </Drawer>
+      </div>
+    );
+  }
+}