+++ app/constants.jsx
@@ -1,0 +1,8 @@
+export const TIMER_START = 'TIMER_START';
+export const TIMER_PAUSE = 'TIMER_PAUSE';
+export const TIMER_RESET = 'TIMER_RESET';
+
+export const POMO_COMPLETE = 'POMO_COMPLETE';
+
+export const BREAK_ADD = 'BREAK_ADD';
+export const BREAK_RESET = 'BREAK_RESET';