+++ app/index.js
@@ -1,0 +1,28 @@
+// @flow
+import injectTapEventPlugin from 'react-tap-event-plugin';
+import React from 'react';
+import { render } from 'react-dom';
+import { Provider } from 'react-redux';
+import { Router, hashHistory } from 'react-router';
+import { syncHistoryWithStore } from 'react-router-redux';
+import darkBaseTheme from 'material-ui/styles/baseThemes/darkBaseTheme';
+import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
+import getMuiTheme from 'material-ui/styles/getMuiTheme';
+
+import routes from './routes';
+import configureStore from './store/configureStore';
+import './reset.global.css';
+import './app.global.css';
+
+const store = configureStore();
+const history = syncHistoryWithStore(hashHistory, store);
+injectTapEventPlugin();
+
+render(
+  <Provider store={store}>
+    <MuiThemeProvider muiTheme={getMuiTheme(darkBaseTheme)}>
+      <Router history={history} routes={routes} />
+    </MuiThemeProvider>
+  </Provider>,
+  document.getElementById('root')
+);