+++ app/pages/NotFound.jsx
@@ -1,0 +1,25 @@
+// Absolute Dependencies
+import React from 'react';
+// Relative Dependencies
+
+// Variable Declarations
+
+class NotFound extends React.Component {
+  render() {
+    return (
+      <div>
+        <p>Page Not Found.</p>
+      </div>
+    );
+  }
+}
+
+NotFound.propTypes = {
+
+};
+
+NotFound.defaultProps = {
+
+};
+
+export default NotFound;