+++ app/reducers/index.js
@@ -1,0 +1,11 @@
+// @flow
+import { combineReducers } from 'redux';
+import { routerReducer as routing } from 'react-router-redux';
+import BreakTrack from './BreakTrack';
+
+const rootReducer = combineReducers({
+  routing,
+  BreakTrack
+});
+
+export default rootReducer;