+++ changelog.md
@@ -1,5 +1,10 @@
 ## BreakTrack Changelog
 
+#### Version: 0.0.2	| 7/11/16
+- Using [electron-react-boilerplate](https://github.com/chentsulin/electron-react-boilerplate)
+- Configured ^ to work with old BreakTrack Code
+- Removed Nodemon
+
 #### Version: 0.0.1 | 6/11/16
 - Added Basic CSS
 - Frameless Window
@@ -10,6 +15,7 @@
 - Readied React Hooks
 
 #### Version: 0.0.0 | 6/11/16
+- (Soft Reset)
 - Started Project
 - Added electron-prebuilt
 - Main Process Built